wpewebkit: Fix ccache usage
oe-core will correctly configure cmake to use ccache based on the CCACHE
bitbake variable, so don't set it explicitly in the wpewebkit recipe.
Setting it explicitly causes issues if a user doesn't want to use ccache
for some reason (for example, if they are using icecream instead).

If using ccache for this specific recipe is desired, one can set the
following in local.conf:

 CCACHE_pn-wpewebkit = "ccache"

Signed-off-by: Joshua Watt <Joshua.Watt@garmin.com>