Update WebKitGTK to 2.28.0 and remove old recipes
* It has been a while we have not updated WebKitGTK recipes.
The last one was 2.22!

* Having all this old recipes here has been a bad idea, as they
are pretty much untested (I bet most of them not build anymore)
and they complicate future upgrades.

* This patch removes all the old recipes and adds a new one for
just the last stable version as of today (2.28.0). The .inc file
its gone and now its merged into the unique .bb recipe

* This patch also adds a new bunch of PACKAGECONFIG options required
to build the last stable version. It also switches the build to Ninja.

Signed-off-by: Carlos Alberto Lopez Perez <clopez@igalia.com>