devupstream: PV ignores the PATCH version and only uses MAJOR.MINOR version
In ./poky/meta/classes/devupstream.bbclass, the PV is defined as follows:

  # Modify the PV if the recipe hasn't already overridden it
  pv = d.getVar("PV")
  proto_marker = "+" + uri.scheme
  if proto_marker not in pv:
      d.setVar("PV", pv + proto_marker + "${SRCPV}")

For preferred viersion like:

  PREFERRED_VERSION_wpewebkit = "2.28.%"

; the devupstream version will be preferred versus the regular version
because 2.28.4+git... > 2.28.4

We skip this issue by setting the PV value to a slightly diffent version
resulting of truncating the minor version number (example: 2.28+git).
The resulting version doesn't match with a three-part version number:

  "2.28+git" doesn't match with "2.28.%".

We can still use "2.28.%" for the release versions and "2.28%" for
devupstream.

Signed-off-by: Pablo Saavedra <psaavedra@igalia.com>