wpebackend-fdo: Add virtual/mesa as dependency
Libs like libwayland-egl should be provided by packages supporting this
virtual package (usually mesa but it could be any other). This library
is normally used for Wayland apps such Weston for stacked Weston servers
and so.

In the top of that, for the case of the wpebackend-fdo makes all the
sense this dependency because this package indeed relies on the FDO
stack.

Signed-off-by: Pablo Saavedra <psaavedra@igalia.com>