wpebackend-fdo: Create wpebackend-fdo_devupstream
We can not use SRCREV to discern for what version is being processed
in the bitbake execution to chose what class inherit. For example:

  inherit ${@bb.utils.contains("SRCREV", 'INVALID', 'cmake', 'meson', d)}

That line is not valid because d.getVar(d, "SRCREV") is evaluated always
to INVALID in this stage.

This patch moves the class-devupstream customizations to a new recipe
(wpebackend-fdo_devupstream.bb). In addition, the 'inherit' statement is
explicitly declared in each recipe.

This fix the regression introduced in
2665a3d557b3c12b891898eb412545ee63aff273 building the bleeding-edge
version.

Unreviewed patch.

Signed-off-by: Pablo Saavedra <psaavedra@igalia.com>