wpewebkit: apply upstream r274943 patch for the 2.30 and 2.32 releases (#247)
Provide and apply the upstream r274943 patch on top of the currently-used
2.30 and 2.32 releases.

This patch adjusts the GStreamer sink creation by including the
imxvideoconvert_g2d element at the beginning of the sink if it's found to be
present on the system. This converter element is needed on some iMX platforms
(generally the ones it's provided on) to perform additional conversion passes
to make the visual output presentable in our pipeline.

It's not a given that the patch will be merged into the 2.30 stable series,
hence it's being added here. It will be included in the next unstable release
so it should be removable whenever the 2.32 version is updated.