qt-wpe-simple-browser: fix runtime and build dependencies
Don't use '+=' since we don't extend those variables.

The wpewebkit package isn't required as build dependency but it is
required as runtime dependency. Without this fix the image containing
the qt-wpe-simple-browser won't be rebuild if wpewebkit is changed.

Finally drop the qtquickcontrols2 build dependency since we don't use
controls in this demo.