wpewebkit: Add libgles2 to runtime dependencies
libgles2 is a runtime dependency of wpewebkit.  Currently libgles2 is
specified as a dependency, but the library isn't directly linked against
by wpewebkit.  This is because wpewebkit uses libepoxy, which loads
libgles2 via dlopen.

Since libgles2 is a runtime dependency of wpewebkit, even though it is
not directly linked against, it is now listed in RDEPENDS.

Signed-off-by: Colin McAllister <colin.mcallister@garmin.com>