Prefer DejaVu instead of Bitstream Vera fonts
Prefer the DejaVu fonts, which are derived from Vera and add better
Unicode coverage, which means they are more suitable as fallback
fonts.

For use as fallback fonts it should be enough to have sans, serif, and
monospace variants. Technically CSS also specifies the cursive and
fantasy generic font families, but in practice they are not used often,
and the sans fallback font will be used if there is none more specific.