cog: Define PACKAGECONFIG options for each version
For those PACKAGECONFIG options which are available only on certain Cog
versions, move them to the version-specific recipe. In order to allow
each recipe to choose default PACKAGECONFIG values, a COG_PACKAGECONFIG
variable can be defined *before* inclusion of "cog.inc", and its value
will be appended to the default PACKAGECONFIG value automatically.