webkitdevci image: optimize list of packages and add quirk for RPi NFS boot.
Some changes to improve the webkitdevci image:

* Optimize the list of packages that go into the image: remove big
unneeded stuff like mariadb/mysql or LTP tests

* Remove also splash: better to see the logs of the init process.

* Add some packages needed to build cog and useful utilities.

* Switch default compression to xz (0 level): is faster than bzip2
and compresses even better.

* Add .tar.xz as a default image format for this image (it will be
the format used in the WebKit CI for the bots) and set WIC to use
xz compression as well.

* Add a quirk for RPi that will copy into the /boot directory all
the files needed to actually boot the RPi. This is very useful to
create an image tarball (format .tar.xz for example) that is ready
to boot via NFS instead of having to deal with individual boot
files not packages inside the tarball.