Rename rdepends from 'virtual/' to 'virtual-'
Recipes shouldn't use "virtual/" in RPROVIDES and RDEPENDS. This was
addressed already in recipes in meta-oe and oe-core. Add a test for
this in insane.bbclass (>nanbield) to ensure no regressions occur.

This change removes all the needless virtuals defined for the RDEPENDS.
It only remains the 'virtual/wpebackend' but renamed now to
'virtual-wpebackend'.

Ref: https://lists.openembedded.org/g/openembedded-core/message/155551