webkitdevi distro: do not disable Weston on systemd.
We are going to start doing tests with Wayland/Weston instead of
DRM/framebuffer so we want it to be available by default.

This change however doesn't make Weston to start by default, so
it remains totally possible to keep testing with DRM/framebuffer
by default if needed.

Weston is started by systemd on demand when the first application
connects to the system-wide socket at WAYLAND_DISPLAY=/run/wayland-0

Signed-off-by: Carlos Alberto Lopez Perez <clopez@igalia.com>