cog_0.16: Fix the launch in Weston kiosk mode
When launching cog in Weston kiosk mode on a board with a display
resolution of 1024 x 600, the following error is seen:

xdg_wm_base@10: error 4: xdg_surface geometry (1024 x 768) is larger than the configured fullscreen state (1024 x 600)

Fix it by backporting an upstream cog commit so that it can be
launched successfully.

This has already been fixed in the cog_0.14.1 recipe. Do the same
for the cog_0.16.1 recipe.

Signed-off-by: Fabio Estevam <festevam@denx.de>