webkit-dev-ci-tools image: ship netdata but with systemd unit disabled
On the WebKit image created via cross-toolchain-helper script we have
been including netdata for a while in the local.conf files there because
we use this to capture performance data from the RPi4 bots when those are
running the tests.

However, this netdata service should not be started by default because
it is only useful on the bots, but it is not useful for developers and
it may cause CPU spikes if not configured properly.

When the bots are deployed we setup a specific configuration for netdata
(to rely the data to a remote server) and we also take care of enabling
the systemd unit there.

So this patch adds netdata into the image but configures it to be disabled
by default. Meanwhile at it, disable also the apache2 service which is
not needed (the service) for the tests (having the binary is enough).